ledger: fix CI-only -Wsign-compare error on GCC/Clang (Linux legs)
test_ledger_model.cpp compared category.Value().value() (an unsigned
long -- the raw FK scalar Light::BelongsTo::Value() returns) directly
against *categoryA/*categoryB (std::int64_t, from CategoryId's
dereference) at three call sites. clang-cl on Windows didn't flag this
under -Weverything, but GCC and Linux Clang's -Wsign-compare (both
under -Werror) correctly caught it -- confirmed as the sole cause of
four failed CI legs (Application ladder, Linux/all-optional-features
on both gcc and clang, Linux/clang-coverage), all failing on the
identical three lines.

Fixed with an explicit static_cast<std::int64_t> on the unsigned side
at each comparison, matching this file's own existing idiom elsewhere
(static_cast<std::int64_t>(row.id.Value()) at every AccountId/LedgerId/
etc. construction site in this same file).

Re-verified locally: ladder_ledger_tests rebuilds clean, 120/120
assertions still pass (unchanged from before this fix -- this was a
warning-level compile issue, not a logic change).

+++ examples/ledger/tests/test_ledger_model.cpp
@@ -392,7 +392,7 @@ TEST_CASE("Replay after editing a rule reproduces the v1 cascade, never the v2 o
                                          *expenseAccountId)
                                   .All();
     REQUIRE(accountRowsBefore.front().category.Value().has_value());
-    CHECK(accountRowsBefore.front().category.Value().value() == *categoryA);
+    CHECK(static_cast<std::int64_t>(accountRowsBefore.front().category.Value().value()) == *categoryA);
 
     // Edit RuleX to v2: now sets Groceries instead of Dining.
     ruleModel.execute(ledger::UpdateRule{.ruleId = ruleId, .matchText = "Coffee", .actionValue = "Groceries"});
@@ -416,8 +416,8 @@ TEST_CASE("Replay after editing a rule reproduces the v1 cascade, never the v2 o
                                         *expenseAccountId)
                                  .All();
     REQUIRE(accountRowsAfter.front().category.Value().has_value());
-    CHECK(accountRowsAfter.front().category.Value().value() == *categoryA);
-    CHECK(accountRowsAfter.front().category.Value().value() != *categoryB);
+    CHECK(static_cast<std::int64_t>(accountRowsAfter.front().category.Value().value()) == *categoryA);
+    CHECK(static_cast<std::int64_t>(accountRowsAfter.front().category.Value().value()) != *categoryB);
 }
 
 TEST_CASE("A clamped Rational leg is caught incidentally by the zero-sum check, not by validate()", "[ledger][rational][security]") {
